fix: properly detect thrown HTTP "error" objects
I'm not sure if minipass-fetch is doing this any longer, or even if it
ever did, as this might be a carry-over from the days of using request
and npm-registry-client.

As far as I can tell, when you get a 4xx or 5xx result, it resolves with
the response, rather than raising a FetchError.  So, this may be dead
code, but it required an update to handle the fact that Nock sets
err.code = 404 when it throws for not having a valid route, and
minipass-fetch copies _all_ the error fields over.