fix: skip integrity check on error
We skip the integrity check of the content when the server has responded
with an error. Otherwise we will hide the real error.

EDIT(@isaacs): This problem was actually addressed in a previous commit,
but the test is good, so I went ahead and pulled it in.  More tests is
good.

PR-URL: https://github.com/npm/make-fetch-happen/pull/29
Credit: @LordCrocket
Close: #29
Reviewed-by: @isaacs