Prune headers that don't belong in cache metadata
There is no justifiable reason to put the `authorization` header in the
cache metadata.  Even if we can assume that the cache should not ever be
exposed or leaked, it's a security footgun that is unnecessary.

Some other fields probably should not be cached as well.  We may want to
re-enable storing the npm or Cloudflare session IDs at some point for
debugging, but for now, these are also just excessive.