feat(BREAKING): refactor cache handling to be more consistent
**BREAKING CHANGE** the `cacheManager` option used to take either a
path, or an instance of an object with specific methods. it has been
deprecated in favor of `cachePath` which is strictly only a path,
and cacache is hard coded as the only caching store. If `cacheManager`
is provided, it will be copied to `cachePath` to ease the transition.

Refactored cache handling to more reliably store and lookup multiple
request/response pairs associated with the same URL

PR-URL: https://github.com/npm/make-fetch-happen/pull/49
Credit: @nlf
Close: #49
Reviewed-by: @wraithgar, @isaacs