fix: always store accept-encoding and content-encoding if present
…regardless of the "compress" option.  Storing them conditionally caused
inconsistency in responses between cache misses vs.  revalidations.
Using them conditionally may have been trying to address underlying
issues posed by not varying on the "compress" option, which I'll change
in a following commit.

Both headers are important for request matching, but are also orthogonal
to the "compress" option.

The accept-encoding header is important even if "compress" is enabled,
as a user-specified accept-encoding may be given to override
minipass-fetch's default.  In any case, it was already being stored,
when present, by the KEEP_REQUEST_HEADERS loop at the top of this
function, regardless of the "compress" option.

Note that default accept-encoding value ("gzip,deflate") sent by
minipass-fetch (if the header is not already set) is still "invisible"
to us on the request object and remains unstored by us.  Such requests
are matched as accept-encoding: identity but sent as accept-encoding:
gzip,deflate, and this may want to be fixed in the future.

The content-encoding header, despite a comment indicating intent
otherwise, was already being stored by the KEEP_RESPONSE_HEADERS loop,
regardless of "compress".  The separate handling of content-encoding
when "compress" was disabled only served to set null values for it,
which could make their way into the response (after stringification) as
content-encoding: null.  This is an invalid value and causes some
clients (like curl) to error.