fix: make `defaults` chaining actually work
The README says:

> A defaulted `fetch` will also have a `.defaults()` method, so they can
> be chained.

But this didn't really work, because while the `defaults` method was
placed on the `defaulted` fetch, a chained defaulted fetch would just
call the original fetch function directly instead of the intermediate
function.

This PR makes chaining actually work and adds a test.

Also, it fixes the previous test, which wasn't actually verifying that
the correct request headers were received because it used the nock API
incorrectly: the option is `reqheaders` not `reqHeaders`.